Move doc to internal
