+++ components/bill/invoice.templ
@@ -6,7 +6,7 @@ import (
 
 	"github.com/invopop/gobl"
 
-	"github.com/invopop/gobl.html/components/doc"
+	"github.com/invopop/gobl.html/internal/doc"
 	"github.com/invopop/gobl.html/components/regimes/co"
 	"github.com/invopop/gobl.html/components/regimes/es"
 	"github.com/invopop/gobl.html/components/regimes/gr"
