+++ components/bill/multipage.templ
@@ -3,7 +3,7 @@ package bill
 import (
 	"github.com/invopop/gobl"
 	"github.com/invopop/gobl.html/components/assets"
-	"github.com/invopop/gobl.html/components/doc"
+	"github.com/invopop/gobl.html/internal/doc"
 	"github.com/invopop/gobl.html/components/t"
 	"github.com/invopop/gobl.html/components/regimes/pt"
 )
