+++ components/regimes/pt/at.templ
@@ -5,7 +5,7 @@ import (
 	"github.com/invopop/gobl/addons/pt/saft"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/tax"
-	"github.com/invopop/gobl.html/components/doc"
+	"github.com/invopop/gobl.html/internal/doc"
 	"github.com/invopop/gobl.html/components/t"
 )
